requirements.txt
Django==1.11.13
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
psycopg2==2.6.2
pytz==2018.4
virtualenv==15.2.0
